requirements.txt
fastapi
databases
motor
pyjwt
email-validator
passlib
aioredis==2.0.0